Create Console app to print remotely received logs
[Changes]
-Created a console app which connects to the logging server via
TCP connection by specified IP and port number.
-The server IP and port number should be specified in command line args
-The logging client listens to messages received from the logging server
and prints out the received info in the console as well as in the logs.txt file.

[Notes]
Needs PR41 for testing.

[Test Cases]
Tested locally... The app successfully connects to the specified server by the
specified port, gets the messages from the logging server and prints in console/file.