[Task 4796]:WinRTC - Include the tests for audio_common in the WinRT solution
[Description]
Include the test for common_audio to WinRT solution.
The tests should be executed via the gtest_runner

[Changes]
-Added common_audio libs and unit test projects to the WinRT solution.
-Added openmax.dl project to WinRT solution to be able to integrate the
common_audio libs and unit tests into the gtest_runner
-Modified the OutputPath() method in fileutils (with a TODO) to be able
to execute test cases which assume a file creation (e.g. .wav files for audio).
This change is required as fopen(), wildly used in the codes, works only for
app local folder for Metro apps.

[Test Cases]
1.Verify that you can separately build the following projects in the WinRT solution:
-common_audio
-common_audio_sse2
-common_audio_unittests
-openmas_dl
2.Verify that you can successfully clean and build the gtest_runner
3.Verify that you can successfully run the gtest_runner
4.Hit the "Run All" button on the gtest_runner app and wait while test are executing
5.At the end the test execution result should be displayed
6.Go to "C:\Users\<user>\AppData\Local\Packages and search for test_detail.xml
7.Open the folder containing the file
8.Verify there are .wav files for audio tests
9.Open the test_detail.xml file and verify that it contains all audio related test
results as well as results of other unit tests.