[Task 4821]:WebRTC - Port webrtc_utility
[Description]
-Build the webrtc_utility project under WinRT (fix compile errors)
-Add the project to the WinRT.sln

[Changes]
-Made the timeGetTime() in trace_win to be a static funtion for WinRT
in order to call it from any other projects
-Replaced the declaration renamed variable name from "generic" to
"genericCodec" in the union as generic is a keyword and cannot be
used as a variable name. Updated the codes were "generic" was accessed.
-Added webrtc_utility into WinRT.sln under "webrtc\modules"

[Test Cases]
1.Verify that the webrtc_utility can successfully built under WinRT
2.Verify that the video_engine_tests project, which has been updated
due to variable name change, also builds properly (not sure if it was
building before, but now it is :) )
3.As a regression, run a ninja build for Win32 and make sure that the
build succeeds and video_engine_tests.exe, video_coding_test.exe and
webrtc_test_common_unittests apps works as before (these are the ones
related to changes).