[Task-5496]: Port video_engine_core
[Description]
Port the video_engine_core to WinRT and integrate video_engine_core_unittests
to gtest_runner

[Changes]
-The video_engine_core_unittests and common_video_unittests prjects are not designed
to be together. To make it possible 2 functions are renamed in the vie_capturer_unittest.cc
to fix link errors
-Added a dummy definition for GetStreamSettings (to be implemented by mosa) to fix
link errors
-Excluded video_capture from gtest_runner (it is not needed at this point and brings in
link errors)
-Declired g_capturePreview in App.cpp to fix link errors

[Notes]
Currently the gtest_runner app fails to build. To test the patch remove the followings
from the Package.appmanifest file:
- xmlns:m3="http://schemas.microsoft.com/appx/2014/manifest" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest"
from the <Package>
and
<mp:PhoneIdentity PhoneProductId="79F175DE-2A85-42D7-9565-55B2020F238E" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>

[Test Cases]
1.Generate projects
2.Make sure you can successfully build and run the gtest_runner
3.verify that all 84 tests from vidoe_engine_core_unittests are passing
4.Verify that all other tests which were passing before are passing now as well