FIX Gtest_runner crash
[Description]
gtest_runner.exe is crashed when launching in build machine or

my local machine.

when debugging, this is a logic error in autodetectproxy.cc, in certain cases, it will start the signalthread twice, which shall not happen.

[test cases]

build the app with VS, and run it with debug, make sure gtest_runner is not crashed, and generated test_details.xml

use ninja build, and launch the appx from command line, make sure it does not crash and generate test_detail.xml