Task 5968:GTest_Runner improvement
[Desription]
add logic so that the gtest_runner could display the current running status.
also add timestamp last time we updated the UI so that we can know if the app is dead or not.

[Test cases]

launch the gtest_runner, make sure very 10seconds, UI is updated to reflect the current test case