[Task 5966]:Fix file access issue on WP8
[Description]
It turned out that the way we copy files in gyp required by unittests
is not working for WP8, as when setting the file's Content it copies
the file right into the app installation folder instead of the desired
folder under the installation folder.

[Changes]
To workaround this issue a resources file has been created in the
libTest-runner project containing the only file required by video tests.
With this hack when you set the file Content it copies the file into the
desired folder.
This is done for Windows Phone only.
Also, enabled the video_coding_test app in the libTest_runner

[Test Cases]
1.Verify that the libTest_runner successfully builds and runs on WP8
and prints out test results (ARM)
2.Verify that the libTest_runner successfully builds and rund on
Windows (Win32)
3.Verify that the ninja build succeeds.