Task 6049:UDP channel creation failure
[root cause analysis]
when creating channel, we copied posix code to validate if a socket descriptor is <FD_SETSIZE.

This logic is not applicable for winrt, as the socket descriptor is not created starting from '0', '1'.
(it starts from a number not used by other app.)

what we really needed is to check if we already opened more sockets than we defined in FD_SETSIZE

Note: The FD_SETSIZE was only used in 'select' posix call for the size of the FD_SET.

In 'UdpSocketManagerWinRTImpl::Process()', we already modifed the posix version's behavior to match with the way how winrt created the socket. When I checked the code there, there is also a related bug there.

[Test case]
launch the startup test app, when 'clicking' start video button, make sure the app does not crash until we can reach

'VideoRenderMediaSourceWinRT::ProcessVideoFrame(const I420VideoFrame& videoFrame)'.

it crashed at  std::memcpy(buffer, videoFrame.buffer(kYPlane), videoFrame.allocated_size(kYPlane) * 3 / 2);

which is a known issue.