[Task 6086]: Port libjingle_media
[Changes]
Modified the gyp file only to fix the compile issues.
sctpdataengine.h and sctpdataengine.cc are not used in WinRT anyhow
at this point, they are hidden in gyp. But there is a TODO in case we
need them.

[Test Cases]
Verify that the libjingle_media project is successfully building under
WinRT