[Task-6143]:Port libjingle_media_unittests to WinRT
[Changes]
-To have the audio related tests pass the voe_base_impl is changed
to use AudioDeviceModule::kWindowsWasapiAudio
-Renamed/added FindFullNameFieldTrialDefault to avoid link conflicts.
libjingle_media_unittests are not designed to be used in a separate app
(e.g. gtest_runner). It uses the field_trial_default lib and the
field_trial lib was already included in the gtest_runner. Both libs have
functions with the same name in the same namespace and it was bringing
link errors. To avoid regressions on other platforms the add/rename of
the function is done for WinRT only.
-Due to the known issue with the mic access the audio device tests are
disabled
-Modified the standup.gyp to avoid link issues due to function rename
change
-Other small fixes for passing the unittests

[Test Cases]
For each step you need to generate projects
-Verify that the gtest_runner, libtest_runner and the standup apps are
successfully builing and running in VS.
-Verify that the ninja build for gtest_runner and libtest_runner builds
successfully
-Verify that the gtest_runner successfully builds and runs on WP