Show current IP in UI
Part of feedback from the initial application demos was that displaying the IP would make things easier.

If the machine is connected to multiple networks, only one IP will be displayed. We've been told this is an unlikely scenario for demos.

The layout is OK on Win 8.1. I expect adjustments will be needed for WP, but I could not test. (the standup project currently fails to build on winRT).