[Issue] after google pull, audio/video delay logging is not working. Previous, we put a temp fix, but that is not correct [Investigation] Found out that the video rtp packet timestamp is not correct.
which is calculated from a frame render time. still need to find out why render time was not set properly.
Put a temp fix which set it to the current time when we have a captured frame, this might artifically shorten the
video delay logging for 1-2ms. will fix it later on.

( Note: In Video_capture_impl.cc, lin 310, there is already a code to set
        _captureFrame.set_ntp_time_ms(captureTime);
        _captureFrame.set_render_time_ms(TickTime::MillisecondTimestamp());

        Still don't know why the it does not set correct timestamp? capturetime is '0'.
        might be that it is not using correct clock? not sure)

 [Test cases]
 make sure we can get audio/video delay in webrtc trace.