Fix Green video screen on phone
[Root cause]
In the last patch of end-to-end audio/video delay logging.

In the QueryOSForTicks(), it reports ticks since epoch.

This is too much for phone, and that code is not optimized.

Thereby, in the phone, decoder process is not synced properly.

[change log]

let QueryOSForTicks() to report 'time' since OS booted (in ms), this would
keep the original intention for QueryOSForTicks().

also, Ticktime:Now(), change it to report the local time since epoch (in ms)

introduced some static variables for TickTime to optimize the code.

[Test case]
make sure PCC calls between phone 8.1/10 and tablet 8.1 behaves normal
make sure end-to-end audio/video logging values are normal