Fix Win32 ninja build
[Changes]
EqualFramesVie() method does not exist anymore.
Replaced its call with EqualFrames() (like in the master branch)

[Note]
There is also one resource file that needs to be copied to the resources
in order to have successful build. I copied it from master branch, however,
in our winrt branch it was not tracked as a change and I could not
add it with this patch. The file name is: google-wifi-3mbps.rx

[Test Cases]
1. Set GYP_GENERATORS to ninja in env variables
2. Generate projects (python webrtc\build\gyp_webrtc)
3. Make a Win32 ninja build: ninja -C out\Debug
4. Verify all projects build
5. Set  GYP_GENERATORS to msvs-winrt in env variables
6. Generate projects (python webrtc\build\gyp_webrtc) for WinRT
for both Windows and Windows phone
7. Make sure video_engine_tests build (it is included neither in gtestRunner
nor in the libTestRunner)