Task 10672: Fix compile errors on Android
[Description]
loggingserver.cc and tracelog.cc were added by us for logging purposes.
However there are lot of functions used in these files which are not
supported for Andoird build.

[Changes]
Modified the GYP file to include these files if we are generating projects
for WinRT only.

[Test Cases]
1. Generate projects for WinRT Windows/Windows Phone
2. Verify that the loggingserver.cc and tracelog.cc are included in the
projects.
3. Generate projects on Android, build with ninja (ninja projects will not
include the loggingserver.cc and tracelog.cc files).
4. Verify you don't get any comple errors.