Added a couple of improvements to the H264 logic.
Added a 60hz timer to ensure we don't reply to sample requests faster than that.
This was necessary because sometimes MF requests samples in burst mode.

Added an MSS starting deferral that wait until there's a frame to render.

Bumped H264 bitrate to 900kbit to improve visual quality.
Explicitly setting constrained baseline on the H264 encoder.