Fix failing unittests
[Changes]
The unittest projects in the webrtc codes are not designed to be
executed in the integrated system. As a result a part of test cases
which belog on a test suite with the same name but located in
different projects were failing.
Renamed some test suites to avoid name conflicts in the
gtest_runner.

[Test Cases]
1. Build and execute the gtest_runner.
2. Make sure there are only 2 failing test cases. You can find the names
of test cases which were failing previously in the zip file attached to the
Task 17619.