Stop collection session as soon as peerconnection instance is destroyed
[NOTE]
The patch should be tested with
https://dpeted.visualstudio.com/DefaultCollection/TED%20Consumer/_git/RTC_Tools/pullrequest/1576?view=discussion

[Description]
The bug "VS Add-in does not show the data in graph for the second call"
can't be fixed with functionality provided by "webrtc_winrt_api". So, we
stop collection session as soon as peerconnection instance is destroyed

[Description of "VS Add-in does not show the data in graph for the second
call"]

PCC metrics data was shown as expected in the first call.
However, after hangup the call, start another call, then, new metrics is
not shown

[Changes]
- add new event type in etw provider to post messages(commands)
- post "stop" command before peerconnection instance is destroyed

[Test Cases]
TC1
1. Follow instructions([DEBUG] section) in OneNote document to Debug the add-in
https://microsoft.sharepoint.com/teams/Real_Time_Communication_for_WinRT/_layouts/15/WopiFrame.aspx?sourcedoc={d0164c70-cb5a-4002-a3d3-91ac6f483d9b}&action=edit&wd=target%28%2F%2FEverything%20you%20need%20to%20start.one%7C60afa497-bbdb-46f3-84fd-d81e11a8b4fb%2FVisual%20Studio%20Add-In%7C3443625f-0811-4a2c-abaf-5ab3a16acfa5%2F%29
2. Establish PCC-PCC call
3. Collect the stats
4. Press "Hang up" or "Disconnect" button

Expected Result: Collection session stops and post-mortem analysis view appears