PR 2207: removed two Todos in Media as no action is needed.
 - removed two Todos in Media as no action is needed.
For explanation, please see commit message.

[Test cases]
(1) for CB, when there is no camera ( for a VM, disable the video device from the VM) or no permission to access to video camera,
when make video call, this result into audio only call, this is desired behaviour

(2) For AppRTC, do the same tests as (1),  also result into audio only call.
also, from logging, observed that video is not part of sdp offer or answer from the testing devices without video.

(3)will report back tests for Jitsimeet and PeerCC